third_party/WebKit/LayoutTests/fullscreen/full-screen-request-rejected-expected.txt

CONSOLE WARNING: Failed to execute 'requestFullScreen' on 'Element': API can only be initiated by a user gesture.
EVENT(webkitfullscreenerror)
END OF TEST
This tests that a full-screen request which is not user-initiated will be rejected.
